Update README.md
Study time

Examine and run the program to see how it functions.

Modify the program so that it also finds the average number of hours per weekend the students study biology as well as programming. Include two prompts, one for each subject. Print out the averages for both subjects.

Also have the program print out which subject the students, on average, spent the most time on.

Note: You can implement the modifications using the existing for loop.